Searchbycategorie to debug

+++ views/pages/article_gestion.php
@@ -13,7 +13,20 @@ session_start();
     // check if user is connected
     require($pathToRootFolder."views/common/checkSessionUser.php");
 
-    
+
+    /// start test debug
+$id1 = htmlspecialchars($_SESSION['id']);
+$id2 = htmlspecialchars($userInfo['id']);
+var_dump($id1);
+var_dump($id2);
+var_dump($userInfo);
+var_dump($_SESSION);
+
+// die();
+/// end test debug
+
+
+
 if(isset($_SESSION['id']) and $userInfo['id'] == $_SESSION['id']) {
     $id = htmlspecialchars($_SESSION['id']);
     $id = htmlspecialchars($userInfo['id']);
@@ -21,9 +34,11 @@ if(isset($_SESSION['id']) and $userInfo['id'] == $_SESSION['id']) {
     // Get my articles
     $myArticles = getMyArticles($id);
     $nbr = 0;
-    foreach($myArticles as $post){
-        if($nbr > 0) {
-            $nbr = $nbr + 1;
+    if(isset($myArticles) AND !empty($myArticles)){
+        foreach($myArticles as $post){
+            if($nbr > 0) {
+                $nbr = $nbr + 1;
+            }
         }
     }
     // $reqnbr = $bdd->prepare('SELECT COUNT(*) FROM comments WHERE author = ?');
@@ -62,14 +77,16 @@ if(isset($_SESSION['id']) and $userInfo['id'] == $_SESSION['id']) {
                             <p class="widgetTextDigit-text">nombre d'articles rédigés</p>
                             <p class="widgetTextDigit-value">
                                 <?php  
-                                $counterNbr = 0;
-                                foreach($myArticles as $myArticle){
-                                    // if($counterNbr >= 0) {
-                                        $counterNbr = $counterNbr + 1;
-                                        
-                                    // }
-                                }
-                                echo $counterNbr;
+                                    $counterNbr = 0;
+                                    if(isset($myArticles) AND !empty($myArticles)){
+                                        foreach($myArticles as $myArticle){
+                                                $counterNbr = $counterNbr + 1;
+                                            }
+                                            echo $counterNbr; 
+                                        }else{
+                                            echo "0";
+                                    }   
+                                
                                 ?>
                             </p>
                         </div>
@@ -83,36 +100,40 @@ if(isset($_SESSION['id']) and $userInfo['id'] == $_SESSION['id']) {
                         </div>
                         <div class="container section-content">
                         <div class="row">
-                            <?php foreach($myArticles as $article): ?>
-                            <?php #for($i=0; $i<6;$i++):?>
-                                <!-- <div class="col-6 show-red"> -->
-                                <!-- </div> -->
+                    <?php   if(isset($myArticles) AND !empty($myArticles)){
+                                    foreach($myArticles as $article): ?>
+                                        <?php #for($i=0; $i<6;$i++):?>
+                                        <!-- <div class="col-6 show-red"> -->
+                                        <!-- </div> -->
 
-                                <div class="blogArticle--medium row   col-lg-6 col-xl-4 px-5 my-5">
-                                    <a class="blogArticle-imglink" >
-                                        <?php foreach($images as $image): ?>
-                                            <?php if($article->id == $image->article_id){ ?>
-                                            <img class="blogArticle-imglink-img" src="../../assets/uploadPersonal/<?php echo $image->name; ?>" alt="image article">
-                                            <?php } ?>
-                                        <?php endforeach; ?>
-                                        
-                                    </a>
-                                    <div class="blogArticle-content">
-                                        <h2 class="blogArticle-title">
-                                            <?= $article->title; ?>
-                                        </h2>
-                                        <span>
-                                            <a href="article_modify.php?edit=<?= $article->id ?>">
-                                                Modifier 
+                                        <div class="blogArticle--medium row   col-lg-6 col-xl-4 px-5 my-5">
+                                            <a class="blogArticle-imglink" >
+                                                <?php foreach($images as $image): ?>
+                                                    <?php if($article->id == $image->article_id){ ?>
+                                                    <img class="blogArticle-imglink-img" src="../../assets/uploadPersonal/<?php echo $image->name; ?>" alt="image article">
+                                                    <?php } ?>
+                                                <?php endforeach; ?>
+                                                
                                             </a>
-                                            <a href="article_delete.php?id=<?= $article->id ?>">
-                                                    <div class="text-danger">Supprimer</div>
-                                            </a>
-                                        </span>
-                                        
-                                    </div>
-                                </div>
-                            <?php endforeach; ?>
+                                            <div class="blogArticle-content">
+                                                <h2 class="blogArticle-title">
+                                                    <?= $article->title; ?>
+                                                </h2>
+                                                <span>
+                                                    <a href="article_modify.php?edit=<?= $article->id ?>">
+                                                        Modifier 
+                                                    </a>
+                                                    <a href="article_delete.php?id=<?= $article->id ?>">
+                                                            <div class="text-danger">Supprimer</div>
+                                                    </a>
+                                                </span>
+                                                
+                                            </div>
+                                        </div>
+                                <?php endforeach; ?>
+                            <?php }else{
+                                echo "Vous n'avez pas rédigé d'articles.</br> <a class=\"dropdown-item\" href:\"article_write.php\">Publier un article </a>\"";
+                            } ?>
 
                         </div>
                     </div>
@@ -132,20 +153,28 @@ if(isset($_SESSION['id']) and $userInfo['id'] == $_SESSION['id']) {
                                 </tr>
                             </thead>
                             <tbody>
-                            <?php foreach($myArticles as $article): ?>
+                            <?php 
+                                if(isset($article) AND !empty($article)){
+                                    if($article->id == $image->article_id){ ?>
+                                    <?php foreach($myArticles as $article): ?>
 
-                                <tr>
-                                    <td class="text-left text-dark"><?= $article->title; ?>.</td>
-                                    <td>
-                                        <a href="article_modify.php?edit=<?= $article->id ?>">
-                                            <i class="fas fa-edit fa-lg text-dominante"></i>
-                                        </a>
-                                    </td>
-                                    <td class="text-secondaire">validé</td>
-                                    <td class="text-secondary"><span class="">12 </span><i class="fas fa-thumbs-up"></i></td>
-                                    <td class="text-secondary"><span class="">22 </span><i class="fas fa-comments"></i></td>
-                                </tr>
-                            <?php endforeach; ?>
+                                        <tr>
+                                            <td class="text-left text-dark"><?= $article->title; ?>.</td>
+                                            <td>
+                                                <a href="article_modify.php?edit=<?= $article->id ?>">
+                                                    <i class="fas fa-edit fa-lg text-dominante"></i>
+                                                </a>
+                                            </td>
+                                            <td class="text-secondaire">validé</td>
+                                            <td class="text-secondary"><span class="">12 </span><i class="fas fa-thumbs-up"></i></td>
+                                            <td class="text-secondary"><span class="">22 </span><i class="fas fa-comments"></i></td>
+                                        </tr>
+                                    <?php endforeach; ?>
+                                    <?php }else{
+                                    echo "Vous n'avez pas rédigé d'articles.</br> <a class=\"dropdown-item\" href:\"article_write.php\">Publier un article </a>\"";
+                                    } 
+                                }
+                            ?>
 
                                 <tr>
                                     <td class="text-left text-dark">Comment changer un carrénage sur Kawa.</td>
@@ -259,7 +288,7 @@ if(isset($_SESSION['id']) and $userInfo['id'] == $_SESSION['id']) {
         <!-- FOOTER -->
         <?php 
             #include($pathToRootFolder."views/common/footer.php");
-            include($pathToRootFolder."views/common/footer_dev_mode.php");
+            #include($pathToRootFolder."views/common/footer_dev_mode.php");
             
         ?>
         
@@ -272,6 +301,7 @@ if(isset($_SESSION['id']) and $userInfo['id'] == $_SESSION['id']) {
 </html>
 <?php
 } else {
+    
     header('Location: session_login.php');
 
 }
